Rename `{EAQ,Batch}Iterable`
